# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  dtos.PaginatedResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/dtos.PaginationMeta'
      success:
        type: boolean
    type: object
  dtos.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dtos.SelectWinnersRequest:
    properties:
      week_number:
        type: integer
    required:
    - week_number
    type: object
...
  dtos.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dtos.WinnerResponse:
    properties:
      created_on:
        type: string
      id:
        type: integer
      qr_code_url:
        type: string
      thunder_seat_id:
        type: integer
      user_id:
        type: string
      week_number:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@thumsup.com
    name: API Support
  description: API documentation for Thums Up Backend Service including Contest Week
    Management
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Thums Up Backend API
  version: "1.0"
paths:
  /admin/winners/select:
    post:
      consumes:
      - application/json
      description: Admin endpoint to select winners for a specific contest week. Requires
        API key authentication.
      parameters:
      - description: Week number
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.SelectWinnersRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Winners selected successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.WinnerResponse'
                  type: array
              type: object
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to select winners
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - APIKey: []
      summary: Select winners for a week
      tags:
      - Admin
...
  /winners:
    get:
      consumes:
      - application/json
      description: Retrieve all winners with pagination support
      parameters:
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        required: true
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Winners retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.WinnerResponse'
                  type: array
              type: object
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to get all winners
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get all winners with pagination
      tags:
      - Winners
  /winners/kyc:
    post:
      consumes:
      - multipart/form-data
      description: After being selected as a winner, user submits their details and
        friends' information.
      parameters:
      - description: User name
        in: formData
        name: user_name
        required: true
        type: string
      - description: User email
        in: formData
        name: user_email
        required: true
        type: string
      - description: Aadhar number
        in: formData
        name: aadhar_number
        required: true
        type: string
      - description: Aadhar front image
        in: formData
        name: aadhar_front
        required: true
        type: file
      - description: Aadhar back image
        in: formData
        name: aadhar_back
        required: true
        type: file
      - description: Friends JSON array
        in: formData
        name: friends
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: KYC submitted successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to submit KYC
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - Bearer: []
      summary: Submit winner KYC details
      tags:
      - Winners
  /winners/week/{weekNumber}:
    get:
      consumes:
      - application/json
      description: Retrieve all winners for a specific contest week
      parameters:
      - description: Week number
        in: path
        name: weekNumber
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Winners retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.WinnerResponse'
                  type: array
              type: object
        "400":
          description: Invalid week number
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to get winners
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get winners by week
      tags:
      - Winners
securityDefinitions:
  APIKey:
    description: Admin API key for protected operations
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"